requirements.txt
sqlalchemy>=2.0.0
click>=8.0.0
tabulate>=0.9.0
python-dateutil>=2.8.0
flask>=3.0.0
requests>=2.31.0
pyllrp>=0.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.0
